Let php-fpm -tt dump info regardless of log_level
When using `php-fpm -tt` to dump the configuration, and the input config has a `log_level` of warning or error, then nothing will be printed.

This change temporarily changes the `zlog` level to notice, if necessary, to allow the output to be printed.

Notably, the "configuration file %s test is successful" output is not affected by this; it still obeys the configured `log_level`.

Can't be tested, however, as `tester.inc`'s log helpers do not actually seem to work with the "invert" flag passed in.

N.b.: the dump probably really should a) not occur with any log decoration applied, and b) be printed to stdout like most other programs do it...