ext/curl: Add all missing CURLE constants up to Curl 8.6
ext/curl now requires libcurl 7.61.0, which means a lot of CURL* constants can be re-organized.

This commit re-arranges all of the CURLE_* constants to a single part of the `curl.stub.php` file,
and rearranges them to match the order of the error constants in the libcurl documentation[^1].

Constants that are deprecated[^2] are moved shifted to the end of the `CURLE_*` constants list.

[^1]: https://curl.se/libcurl/c/libcurl-errors.html
[^2]: https://curl.se/libcurl/c/symbols-in-versions.html

Related: #12000, #13259, #13209, #13282