Remove PHP atomic includes and PHP_DEFINE M4 macro
PHP_DEFINE was introduced with the PHP 5 build system
9d9d39a0de3bec962c343051011f5a2ed7d7b242 and then refactored via
350de12bc24472ccf20ff9a8b39a6da0185c070d.

This was once used to put defined constants into a single file to have
more fine-graned dependencies (atomic includes). Since no known PHP
extension is using this and it makes very little sense to use this, this
M4 macro can be removed in favor of the Autoconf native way using
AC_DEFINE and the usual included files php_config.h and config.h.

- Generated unused include directory removed
- Remove include dir from DEFS
- Remove also include dir from PDO checks