Minor PDO conditions cleanup (#13584)
* Get rid of inverted NULL checks

As the dbh pointer is already used in address computation,
the NULL checks is redundant. Looking at the callers
there are no cases where this can be passed as NULL.

* Message can't be NULL here

* Remove null check for return_value

As one of the branches dereferences return_value unconditionally, the
NULL check is dubious. In fact, looking at the callers we see that
return_value can never be NULL. So remove the check and fix the comment.