CI: Cirrus CI - skip build on ignore-paths
Cirrus CI supports a `<job>.skip` key that skips the build if the expression
evaluates to true. This adds the same list of `on.pull_request.paths-ignore`
patterns from GitHub Actions workflows.

This should save several minutes of CI times on PRs when the changes are
only in the README/doc files.

[Cirrus CI Docs](https://cirrus-ci.org/guide/writing-tasks/#supported-functions)