Fix posix_spawn cwd incompatibility
We switched from fork+exec to posix_spawn for proc_open when available in PHP
8.3. Previously, a cwd that does not exist, or doesn't have the executable chmod
flag set would silent be ignored. posix_spawn fails, on the other hand.

We provide better backwards compatibility by manually checking for the path.
This patch is technically subject to data races. I'm not sure there is a way to
avoid this with posix_spawn.

Fixes GH-13743