Adjust GC threshold if num_roots higher than threshold after collection
This fixes an edge causing the GC to be triggered repeatedly.

Destructors might add potential garbage to the buffer, so it may happen that
num_root it higher than threshold after collection, thus triggering a GC run
almost immediately. This can happen by touching enough objects in a destructor,
e.g. by iterating over an array. If the happens again in the new run, and the
threshold is not updated, the GC may be triggered repeatedly.

Also prevent GC runs during zend_gc_check_root_tmpvars(), as the threshold is
not adjusted yet a this point.